Move around StringWrapper and co, so only one declaration/definition exist
