+++ profiling-ffi/cbindgen.toml
@@ -47,6 +47,8 @@ renaming_overrides_prefixing = true
 "Slice_File" = "ddog_prof_Exporter_Slice_File"
 "ManagedStringStorage" = "ddog_prof_ManagedStringStorage"
 "ManagedStringId" = "ddog_prof_ManagedStringId"
+"StringWrapper" = "ddog_StringWrapper"
+"StringWrapperResult" = "ddog_StringWrapperResult"
 
 [export.mangle]
 rename_types = "PascalCase"
